+++ Linux_System_RT/time_mask.c
@@ -92,7 +92,7 @@ struct pdw_s *emitter_to_pdws(struct emitter_s *em, int num_emitters, double us_
 		pdws[j] = malloc(sizeof(struct pdw_s) * num_pdws);
 		Size[j] = num_pdws;
 		for(int i = 0; i<Size[j]; i++){
-			uint64_t TOA = T0 + (unsigned int)em[j].offset + i*em[j].PRI;
+			double TOA = T0 + (double)em[j].offset + i*em[j].PRI;
 			//printf("TOA: %d\n", TOA);
 			struct pdw_s pdw = {TOA, em[j].MOP, false, false, false, em[j].FREQ_OFFSET,em[j].LEVEL_OFFSET,em[j].PHASE_OFFSET, em[j].EDGE_TYPE, em[j].SEGMENT_IDX, em[j].PW, em[j].FREQ_INC, em[j].CHIP_WIDTH, em[j].CODE, em[j].RISE_TIME, em[j].FALL_TIME, false, 0, 0};
 			pdws[j][i] = pdw;
